change setup script

+++ start.sh
@@ -1,6 +1,7 @@
 #!/bin/sh
 
 project_name=agent-ios
+bundle_path=vendor/bundle
 
 separator() {
     echo "-------------"
@@ -13,7 +14,9 @@ invoke() {
 }
 
 ## install gem cocoapods
-invoke "bundle install --path vendor/bundle"
+if [ ! -e $bundle_path ]; then
+    invoke "bundle install --path vendor/bundle"
+fi
 
 ## install pods
 invoke "bundle exec pod install"
